requirements.txt
firebase-admin~=6.6.0
cryptography~=44.0.1
python-dotenv~=1.0.1
fastapi==0.68.0
uvicorn==0.15.0
tensorflow==2.16.1
Pillow==8.3.1
python-multipart
numpy>=1.19.2